[Update] Update save image url

+++ IngredientServer/Core/Services/IngredientService.cs
@@ -2,32 +2,42 @@
 using IngredientServer.Core.Interfaces.Repositories;
 using IngredientServer.Core.Interfaces.Services;
 using IngredientServer.Utils.DTOs.Entity;
+using Microsoft.AspNetCore.Http;
 
 namespace IngredientServer.Core.Services;
 
-public class IngredientService(IIngredientRepository ingredientRepository, IUserContextService userContextService)
+public class IngredientService(IIngredientRepository ingredientRepository, IUserContextService userContextService,IImageService imageService)
     : IIngredientService
 {
-    public async Task<IngredientDto> CreateIngredientAsync(IngredientDataDto dataDto)
+    public async Task<IngredientDataResponseDto> CreateIngredientAsync(CreateIngredientRequestDto dto)
     {
+        //Save Image 
+        string? imageUrl = null;
+        if (dto.Image != null && dto.Image.Length > 0)
+        {
+            imageUrl = await imageService.SaveImageAsync(dto.Image);
+        }
+        
         var ingredient = new Ingredient
         {
-            Name = dataDto.Name,
-            Description = dataDto.Description,
-            Category = dataDto.Category,
-            Quantity = dataDto.Quantity,
-            ExpiryDate = dataDto.ExpiryDate,
-            Unit = dataDto.Unit,
-            UserId = userContextService.GetAuthenticatedUserId()
+            Name = dto.Name,
+            Description = dto.Description,
+            Category = dto.Category,
+            Quantity = dto.Quantity,
+            ExpiryDate = dto.ExpiryDate,
+            Unit = dto.Unit,
+            UserId = userContextService.GetAuthenticatedUserId(),
+            ImageUrl = imageUrl
         };
 
         var savedIngredient = await ingredientRepository.AddAsync(ingredient);
+        
         if (savedIngredient == null)
         {
             throw new UnauthorizedAccessException("Failed to create ingredient or access denied.");
         }
         // Map the saved ingredient to DTO
-        return new IngredientDto
+        return new IngredientDataResponseDto
         {
             Id = savedIngredient.Id,
             Name = savedIngredient.Name,
@@ -36,12 +46,19 @@ public class IngredientService(IIngredientRepository ingredientRepository, IUser
             Category = savedIngredient.Category,
             Quantity = savedIngredient.Quantity,
             ExpiryDate = savedIngredient.ExpiryDate,
+            ImageUrl = savedIngredient.ImageUrl
         };
     }
 
-    public async Task<IngredientDto> UpdateIngredientAsync(int ingredientId, IngredientDataDto dto)
+    public async Task<IngredientDataResponseDto> UpdateIngredientAsync(UpdateIngredientRequestDto dto)
     {
-        var ingredient = await ingredientRepository.GetByIdAsync(ingredientId);
+        if (dto.Id <= 0)
+        {
+            throw new ArgumentException("Invalid ingredient ID", nameof(dto.Id));
+        }
+        
+        var ingredient = await ingredientRepository.GetByIdAsync(dto.Id);
+        
         if (ingredient == null)
         {
             throw new UnauthorizedAccessException("Ingredient not found or access denied.");
@@ -53,9 +70,21 @@ public class IngredientService(IIngredientRepository ingredientRepository, IUser
         ingredient.Category = dto.Category;
         ingredient.Quantity = dto.Quantity;
         ingredient.ExpiryDate = dto.ExpiryDate;
+        
+        if (dto.Image is { Length: > 0 })
+        {
+            if (string.IsNullOrEmpty(ingredient.ImageUrl))
+            {
+                ingredient.ImageUrl = await imageService.SaveImageAsync(dto.Image);
+            }
+            else
+            {
+                ingredient.ImageUrl = await imageService.UpdateImageAsync(dto.Image, ingredient.ImageUrl);
+            }
+        }
 
         var updatedIngredient = await ingredientRepository.UpdateAsync(ingredient);
-        return new IngredientDto
+        return new IngredientDataResponseDto
         {
             Id = updatedIngredient.Id,
             Name = updatedIngredient.Name,
@@ -64,13 +93,29 @@ public class IngredientService(IIngredientRepository ingredientRepository, IUser
             Category = updatedIngredient.Category,
             Quantity = updatedIngredient.Quantity,
             ExpiryDate = updatedIngredient.ExpiryDate,
+            ImageUrl = updatedIngredient.ImageUrl
         };
     }
 
     public async Task<bool> DeleteIngredientAsync(int ingredientId)
     {
+        if (ingredientId <= 0)
+        {
+            throw new ArgumentException("Invalid ingredient ID", nameof(ingredientId));
+        }
+
+        var ingredient = await ingredientRepository.GetByIdAsync(ingredientId);
+
+        if (ingredient == null)
+        {
+            throw new UnauthorizedAccessException("Ingredient not found or access denied.");
+        }
+
+        if (string.IsNullOrEmpty(ingredient.ImageUrl)) return await ingredientRepository.DeleteAsync(ingredientId);
+        await imageService.DeleteImageAsync(ingredient.ImageUrl);
         return await ingredientRepository.DeleteAsync(ingredientId);
     }
+
 
     public async Task<IngredientSearchResultDto> GetAllAsync(IngredientFilterDto filter)
     {
@@ -78,14 +123,14 @@ public class IngredientService(IIngredientRepository ingredientRepository, IUser
         return ingredients;
     }
 
-    public async Task<IngredientDto> GetIngredientByIdAsync(int id)
+    public async Task<IngredientDataResponseDto> GetIngredientByIdAsync(int id)
     {
         var ingredient = await ingredientRepository.GetByIdAsync(id);
         if (ingredient == null)
         {
             throw new UnauthorizedAccessException("Ingredient not found or access denied.");
         }
-        return new IngredientDto
+        return new IngredientDataResponseDto
         {
             Id = ingredient.Id,
             Name = ingredient.Name,
@@ -94,9 +139,7 @@ public class IngredientService(IIngredientRepository ingredientRepository, IUser
             Category = ingredient.Category,
             Quantity = ingredient.Quantity,
             ExpiryDate = ingredient.ExpiryDate,
-            DaysUntilExpiry = (ingredient.ExpiryDate - DateTime.UtcNow).Days,
-            IsExpired = ingredient.ExpiryDate < DateTime.UtcNow,
-            IsExpiringSoon = (ingredient.ExpiryDate - DateTime.UtcNow).Days <= 7
+            ImageUrl = ingredient.ImageUrl
         };
     }
 }
